started to implement serializer using standard protobuf
